+++ src/app/app.module.ts
@@ -16,6 +16,7 @@ import { MainComponent } from './main/main.component';
 import { WordItemComponent } from './word-item/word-item.component';
 import { MenuButtonComponent } from './menu-button/menu-button.component';
 import { YesNoDialogComponent } from './yes-no-dialog/yes-no-dialog.component';
+import { ButtonMenuSelectComponent } from './button-menu-select/button-menu-select.component';
 
 
 @NgModule({
@@ -24,7 +25,8 @@ import { YesNoDialogComponent } from './yes-no-dialog/yes-no-dialog.component';
     MainComponent,
     WordItemComponent,
     MenuButtonComponent,
-    YesNoDialogComponent
+    YesNoDialogComponent,
+    ButtonMenuSelectComponent
   ],
   imports: [
     BrowserModule,