+++ src/app/class/voice-item.spec.ts
@@ -1,0 +1,7 @@
+import { VoiceItem } from './voice-item';
+
+describe('VoiceItem', () => {
+  it('should create an instance', () => {
+    expect(new VoiceItem()).toBeTruthy();
+  });
+});