+++ src/app/class/voice-item.ts
@@ -1,0 +1,10 @@
+export class VoiceItem {
+    name: string;
+    text: string;
+  
+    constructor(name: string = "", text: string = "") {
+      this.name = name;
+      this.text = text;
+    }
+  }
+