+++ src/app/class/word-for-view.spec.ts
@@ -1,6 +1,6 @@
 import { WordForView } from './word-for-view';
 
-describe('VocabularyViewItem', () => {
+describe('WordForView', () => {
   it('should create an instance', () => {
     expect(new WordForView()).toBeTruthy();
   });