+++ src/app/class/word.spec.ts
@@ -1,6 +1,6 @@
 import { Word } from './word';
 
-describe('VocabularyItem', () => {
+describe('Word', () => {
   it('should create an instance', () => {
     expect(new Word()).toBeTruthy();
   });